Use single floats instead of doubles in box

+++ include/compiler/vm/box.hpp
@@ -12,7 +12,7 @@ namespace salmon::vm {
 	struct Box : public AllocatedItem {
 		~Box();
 		struct Type *type;
-		std::variant<int32_t, double,salmon::vm::vm_ptr<Symbol>,
+		std::variant<int32_t, float,salmon::vm::vm_ptr<Symbol>,
 			salmon::vm::vm_ptr<StaticString>> elem;
 	};
 }
